+++ src/Rules/RequiredWithout.php
@@ -2,8 +2,12 @@
 
 namespace Somnambulist\Components\Validation\Rules;
 
+use Somnambulist\Components\Validation\Rules\Behaviours\CanConvertFieldParametersDotNotationToResolvedStrings;
+
 class RequiredWithout extends Required
 {
+    use CanConvertFieldParametersDotNotationToResolvedStrings;
+
     protected bool $implicit = true;
     protected string $message = 'rule.required_without';
 
@@ -22,7 +26,7 @@ class RequiredWithout extends Required
         $requiredValidator = $this->validation->factory()->rule('required');
 
         foreach ($fields as $field) {
-            if (!$this->validation->input()->has($field)) {
+            if (!$this->attribute->value($field)) {
                 $this->setAttributeAsRequired();
 
                 return $requiredValidator->check($value);