+++ tests/Rules/BeforeTest.php
@@ -3,6 +3,7 @@
 namespace Somnambulist\Components\Validation\Tests\Rules;
 
 use DateTime;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Somnambulist\Components\Validation\Exceptions\ParameterException;
 use Somnambulist\Components\Validation\Rules\Before;
@@ -20,9 +21,7 @@ class BeforeTest extends TestCase
         $this->validator = new Before();
     }
 
-    /**
-     * @dataProvider getValidDates
-     */
+    #[DataProvider('getValidDates')]
     public function testOnlyAWellFormedDateCanBeValidated($date)
     {
         $this->assertTrue(
@@ -44,9 +43,7 @@ class BeforeTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider getInvalidDates
-     */
+    #[DataProvider('getInvalidDates')]
     public function testANonWellFormedDateCannotBeValidated($date)
     {
         $this->expectException(ParameterException::class);